docs: record the WmlComparer removal, and fix the npm tests it invalidated
CHANGELOG gets the breaking entries, written for someone upgrading: what is gone,
what moved, and what silently changes for a caller who passes nothing. The last
point matters most -- DocxCompare.Compare still applies PreAcceptInputRevisions and
PreserveInputRevisions, so moving a call to DocxDiff.Compare without setting both
produces whole-document churn on revision-bearing inputs.

CLAUDE.md's numbers were stale the moment the engine went: library 135 -> 133
warnings, tests 779 -> 713, suite ~4,260 -> ~3,900, and the all-21-stripped nullable
figure 3,659 -> 2,977. The count of #nullable-disable files stays at 21 because
PackageMerge.cs replaces WmlComparer.cs in that set. Its "DocxDiff still depends on
WmlComparer's types, so the older engine is not removable" note was wrong on the
detail (DocxDiff referenced those types only in doc comments; the real dependency
was three package-merge helpers) and is replaced by what a reader now needs: that
DocxCompare and DocxDiff differ by the input-revision policy.

The four WmlComparer design docs are kept and banner-marked historical rather than
deleted -- they record Word-behaviour findings DocxDiff inherited or deliberately
diverged from, which would be expensive to rediscover.

npm tests: engine-selector.spec.ts swept a selector that no longer exists, but two
of its assertions were about the comparison itself, so it becomes
compare-front-door.spec.ts keeping those. The trim-validation spec still guards the
same reflective GetPackage() path -- that path is now PackageMerge, reached from
DocxDiff's renderer. The harness's getRevisions unwrapped a { Revisions: [...] }
envelope that is now a bare array, and docxodus.spec.ts asserted every revision
carries moveGroupId/isMoveSource; it now checks the id/family/resolutionStatus the
grouped entries actually have.

Verified: npm tsc --noEmit clean, npm run build clean (WASM wire size 3656 KB
against a 4096 KB budget).
