+++ docs/architecture/comparison_engine.md
@@ -1,5 +1,11 @@
 # DOCX Comparison and HTML Rendering Engine
 
+> **Historical (v11.0.0).** This document describes `WmlComparer`, the legacy comparison engine,
+> which was removed in v11.0.0. It is kept because it records design decisions and Word-behaviour
+> findings that `DocxDiff` inherited or deliberately diverged from — not because the code still
+> exists. For the shipping engine see `ir_diff_engine.md`; for the parity evidence gathered before
+> removal see `wmlcomparer_parity_baseline/`.
+
 This document describes the architecture of the document comparison and tracked changes rendering system in Docxodus, covering how two Word documents are compared to produce tracked changes, and how those changes can be rendered to HTML.
 
 ## Overview
